+++ minify-html-onepass/src/unit/comment.rs
@@ -4,14 +4,12 @@ use crate::proc::MatchMode::*;
 use crate::proc::Processor;
 use aho_corasick::AhoCorasick;
 use aho_corasick::AhoCorasickBuilder;
-use aho_corasick::AhoCorasickKind;
 use once_cell::sync::Lazy;
 
 static COMMENT_END: Lazy<AhoCorasick> = Lazy::new(|| {
   AhoCorasickBuilder::new()
-    .kind(Some(AhoCorasickKind::DFA))
+    .dfa(true)
     .build(["-->"])
-    .unwrap()
 });
 
 #[inline(always)]