+++ minify-html-onepass/src/unit/style.rs
@@ -5,7 +5,6 @@ use crate::proc::Processor;
 use crate::Cfg;
 use aho_corasick::AhoCorasick;
 use aho_corasick::AhoCorasickBuilder;
-use aho_corasick::AhoCorasickKind;
 use lightningcss::stylesheet::MinifyOptions;
 use lightningcss::stylesheet::ParserOptions;
 use lightningcss::stylesheet::PrinterOptions;
@@ -16,9 +15,8 @@ use std::str::from_utf8_unchecked;
 static STYLE_END: Lazy<AhoCorasick> = Lazy::new(|| {
   AhoCorasickBuilder::new()
     .ascii_case_insensitive(true)
-    .kind(Some(AhoCorasickKind::DFA))
+    .dfa(true)
     .build(["</style"])
-    .unwrap()
 });
 
 #[inline(always)]