+++ minify-html/src/minify/attr.rs
@@ -1,7 +1,6 @@
 use crate::entity::encode::encode_entities;
 use crate::Cfg;
 use aho_corasick::AhoCorasickBuilder;
-use aho_corasick::AhoCorasickKind;
 use aho_corasick::MatchKind;
 use lightningcss::stylesheet::MinifyOptions;
 use lightningcss::stylesheet::ParserOptions;
@@ -34,10 +33,9 @@ fn build_double_quoted_replacer() -> Replacer {
 
   Replacer::new(
     AhoCorasickBuilder::new()
-      .kind(Some(AhoCorasickKind::DFA))
+      .dfa(true)
       .match_kind(MatchKind::LeftmostLongest)
-      .build(patterns)
-      .unwrap(),
+      .build(patterns),
     replacements,
   )
 }
@@ -46,9 +44,8 @@ fn build_double_quoted_replacer() -> Replacer {
 fn build_whatwg_double_quoted_replacer() -> Replacer {
   Replacer::new(
     AhoCorasickBuilder::new()
-      .kind(Some(AhoCorasickKind::DFA))
-      .build([b"\""])
-      .unwrap(),
+      .dfa(true)
+      .build([b"\""]),
     vec![b"&#34;".to_vec()],
   )
 }
@@ -68,10 +65,9 @@ fn build_single_quoted_replacer() -> Replacer {
 
   Replacer::new(
     AhoCorasickBuilder::new()
-      .kind(Some(AhoCorasickKind::DFA))
+      .dfa(true)
       .match_kind(MatchKind::LeftmostLongest)
-      .build(patterns)
-      .unwrap(),
+      .build(patterns),
     replacements,
   )
 }
@@ -80,9 +76,8 @@ fn build_single_quoted_replacer() -> Replacer {
 fn build_whatwg_single_quoted_replacer() -> Replacer {
   Replacer::new(
     AhoCorasickBuilder::new()
-      .kind(Some(AhoCorasickKind::DFA))
-      .build([b"'"])
-      .unwrap(),
+      .dfa(true)
+      .build([b"'"]),
     vec![b"&#39;".to_vec()],
   )
 }
@@ -128,10 +123,9 @@ fn build_unquoted_replacer() -> Replacer {
 
   Replacer::new(
     AhoCorasickBuilder::new()
-      .kind(Some(AhoCorasickKind::DFA))
+      .dfa(true)
       .match_kind(MatchKind::LeftmostLongest)
-      .build(patterns)
-      .unwrap(),
+      .build(patterns),
     replacements,
   )
 }
@@ -204,10 +198,9 @@ fn build_semi_whatwg_unquoted_replacer() -> Replacer {
 
   Replacer::new(
     AhoCorasickBuilder::new()
-      .kind(Some(AhoCorasickKind::DFA))
+      .dfa(true)
       .match_kind(MatchKind::LeftmostLongest)
-      .build(patterns)
-      .unwrap(),
+      .build(patterns),
     replacements,
   )
 }
@@ -247,10 +240,9 @@ fn build_whatwg_unquoted_replacer() -> Replacer {
 
   Replacer::new(
     AhoCorasickBuilder::new()
-      .kind(Some(AhoCorasickKind::DFA))
+      .dfa(true)
       .match_kind(MatchKind::LeftmostLongest)
-      .build(patterns)
-      .unwrap(),
+      .build(patterns),
     replacements,
   )
 }