+++ minify-html/src/minify/content.rs
@@ -11,7 +11,6 @@ use crate::minify::element::minify_element;
 use crate::minify::instruction::minify_instruction;
 use crate::minify::js::minify_js;
 use aho_corasick::AhoCorasickBuilder;
-use aho_corasick::AhoCorasickKind;
 use aho_corasick::MatchKind;
 use minify_html_common::gen::codepoints::TAG_NAME_CHAR;
 use minify_html_common::pattern::Replacer;
@@ -39,10 +38,9 @@ fn build_optimal_chevron_replacer() -> Replacer {
 
   Replacer::new(
     AhoCorasickBuilder::new()
-      .kind(Some(AhoCorasickKind::DFA))
+      .dfa(true)
       .match_kind(MatchKind::LeftmostLongest)
-      .build(patterns)
-      .unwrap(),
+      .build(patterns),
     replacements,
   )
 }
@@ -50,9 +48,8 @@ fn build_optimal_chevron_replacer() -> Replacer {
 fn build_whatwg_chevron_replacer() -> Replacer {
   Replacer::new(
     AhoCorasickBuilder::new()
-      .kind(Some(AhoCorasickKind::DFA))
-      .build(["<"])
-      .unwrap(),
+      .dfa(true)
+      .build(["<"]),
     vec!["&lt;".into()],
   )
 }