+++ minify-html/src/parse/instruction.rs
@@ -2,14 +2,12 @@ use crate::ast::NodeData;
 use crate::parse::Code;
 use aho_corasick::AhoCorasick;
 use aho_corasick::AhoCorasickBuilder;
-use aho_corasick::AhoCorasickKind;
 use once_cell::sync::Lazy;
 
 static INSTRUCTION_END: Lazy<AhoCorasick> = Lazy::new(|| {
   AhoCorasickBuilder::new()
-    .kind(Some(AhoCorasickKind::DFA))
+    .dfa(true)
     .build(["?>"])
-    .unwrap()
 });
 
 pub fn parse_instruction(code: &mut Code) -> NodeData {