+++ minify-html/src/parse/script.rs
@@ -4,15 +4,13 @@ use crate::parse::content::ParsedContent;
 use crate::parse::Code;
 use aho_corasick::AhoCorasick;
 use aho_corasick::AhoCorasickBuilder;
-use aho_corasick::AhoCorasickKind;
 use once_cell::sync::Lazy;
 
 static END: Lazy<AhoCorasick> = Lazy::new(|| {
   AhoCorasickBuilder::new()
     .ascii_case_insensitive(true)
-    .kind(Some(AhoCorasickKind::DFA))
+    .dfa(true)
     .build(["</script"])
-    .unwrap()
 });
 
 pub fn parse_script_content(code: &mut Code, lang: ScriptOrStyleLang) -> ParsedContent {